review fixes

+++ eth/stagedsync/bor_heimdall_shared.go
@@ -86,7 +86,7 @@ func fetchRequiredHeimdallSpansIfNeeded(
 		}
 	}
 
-	return uint64(requiredSpanID), err
+	return requiredSpanID, err
 }
 
 func fetchAndWriteHeimdallSpan(
